fix: refactor more
