+++ frontend/app/src/pages/main/v1/workflow-runs-v1/$run/v2components/workflow-run-visualizer-v2.tsx
@@ -10,9 +10,9 @@ import 'reactflow/dist/style.css';
 import dagre from 'dagre';
 import { useTheme } from '@/components/theme-provider';
 import stepRunNode, { NodeData } from './step-run-node';
-import { useWorkflowDetails } from '../../hooks';
 import invariant from 'tiny-invariant';
 import { V1TaskStatus } from '@/lib/api';
+import { useWorkflowDetails } from '../../hooks/use-workflow-details';
 
 const connectionLineStyleDark = { stroke: '#fff' };
 const connectionLineStyleLight = { stroke: '#000' };
